Delete specialized Monitor files
