+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Threading/ObjectHeader.NativeAot.cs
