+++ src/coreclr/vm/ecalllist.h
@@ -260,8 +260,8 @@ FCFuncStart(gThreadFuncs)
     FCFuncElement("get_OptimalMaxSpinWaitsPerSpinIteration", ThreadNative::GetOptimalMaxSpinWaitsPerSpinIteration)
 FCFuncEnd()
 
-FCFuncStart(gMonitorFuncs)
-    FCFuncElement("GetLockHandleIfExists", Monitor_GetLockHandleIfExists)
+FCFuncStart(gObjectHeaderFuncs)
+    FCFuncElement("GetLockHandleIfExists", ObjectHeader_GetLockHandleIfExists)
 FCFuncEnd()
 
 
@@ -402,7 +402,7 @@ FCClassElement("Math", "System", gMathFuncs)
 FCClassElement("MathF", "System", gMathFFuncs)
 FCClassElement("MetadataImport", "System.Reflection", gMetaDataImport)
 FCClassElement("MethodTable", "System.Runtime.CompilerServices", gMethodTableFuncs)
-FCClassElement("Monitor", "System.Threading", gMonitorFuncs)
+FCClassElement("ObjectHeader", "System.Threading", gObjectHeaderFuncs)
 FCClassElement("RuntimeAssembly", "System.Reflection", gRuntimeAssemblyFuncs)
 FCClassElement("RuntimeFieldHandle", "System", gCOMFieldHandleNewFuncs)
 FCClassElement("RuntimeHelpers", "System.Runtime.CompilerServices", gRuntimeHelpers)
